test: create separeted file for exact match evaluation test
